requirements.txt
requests==2.32.3
google-generativeai==0.8.4
pandas==2.2.0
flask==3.0.3
flask-cors==4.0.0
jupyter==1.0.0
notebook==7.0.7 